Add namepace for Manager
