+++ app/controllers/admin/base_controller.rb
@@ -3,7 +3,7 @@ module Admin
     before_action :restrict_non_admin
 
     def restrict_non_admin
-      #return redirect_to root_path unless current_user.admin?
+      #return redirect_to root_path unless current_user.admin? 
     end
   end
 end
