+++ app/controllers/manager/comments_controller.rb
@@ -1,4 +1,4 @@
-module Admin
+module Manager
   class CommentsController < BaseController
     before_action :authenticate_user!
     before_action :set_project
