+++ app/views/manager/clients/_form.html.erb
