+++ app/views/manager/clients/index.html.erb
