+++ app/views/manager/clients/new.html.erb
