+++ app/views/manager/clients/show.html.erb
