+++ app/views/manager/comments/create.js.erb
