+++ app/views/manager/projects/_form.html.erb
