+++ app/views/manager/projects/index.html.erb
