+++ app/views/manager/projects/new.html.erb
