+++ app/views/manager/projects/show.html.erb
