+++ config/routes.rb
@@ -2,15 +2,17 @@ Rails.application.routes.draw do
 
   devise_for :users
 
-  namespace :admin do
-    resources :users
+  namespace :manager do
+    resources :clients
     resources :projects do 
       resources :comments
     end
   end 
 
+  namespace :admin do
+    resources :users
+  end
+
   root "admin/projects#index"
-  
-  resources :clients
 
 end
